Uppdaterat alla delete-skript.

+++ Hemsida/PHP/DeleteForetag.php
@@ -5,17 +5,23 @@ $password="2ykgB03hnx"; // Mysql password
 $db_name="sg_systemet_com"; // Database name 
 $tbl_name="SG_Foretag"; // Table name 
 
+// Anger att det är text som skall produceras. 
+header('Content-type: text/plain');
+
 // Connect to server and select databse.
-mysql_connect("$host", "$username", "$password")or die("cannot connect"); 
-mysql_select_db("$db_name")or die("cannot select DB");
+$con = mysqli_connect($host, $username, $password, $db_name);
+if ($mysqli->connect_errno) {
+    printf("Anslutningsfel: %s\n", $mysqli->connect_error);
+    exit();
+}
 
 // Data sent from form .NET
 $OrderID=$_POST['OrderID'];
 
 // Delete
 $sql="DELETE FROM `$tbl_name` WHERE OrderID = '$OrderID'";
-$result=mysql_query($sql);
-$affectedRows = mysql_affected_rows();
+$result = $con->query($sql);
+$rc = $con->affected_rows;
 
 // Checks if success
 echo "<?xml version=\"1.0\" encoding=\"UTF-8\"?>";
@@ -37,5 +43,6 @@ echo "</Data>";
 echo "\n";
 echo "</MessageXML>";
 
-mysql_close();
+$result->free();
+$con->close();
 ?>
